PullViewDisplayMode: pull views in Scroller may be fixed or dragged
Previously, pull views could only be dragged. if a pull view is fixed, it will
always be positioned exactly at the edge of the container. if hasElasticEdges is
true, then the pull view will be revealed from under the scroller's view port.
if hasElasticEdges is false, the pull view will be made visible on top of the
scroller's view port. (closes #1514)