TextFieldTextRenderer: fix last line getting lost (Feathers 2.3)
Feathers 2.x has a bug with TextFieldTextRenderer, where the last line of a multi-line Label sometimes is not rendered. This occurs when the calculated height of the text is a fraction. This issue is resolved by rounding up the height to the nearest integer.

Although Feathers 2.3 is no longer in active development, this fix is critical for legacy products which rely on Feathers 2.3.