FontStylesSet: added dispose() method and no longer clones TextFormat objects (because it needs to listen for Event.CHANGE from the original format).
Components that uses FontStylesSet need to dispose it when they are disposed. This ensures that events are removed to avoid the memory leak that the clone() call was working around. I also added a bunch of new tests for text renderers/editors and FontStylesSet to ensure that this behavior says consistent.